ref: simplify test
